+++ Changes.md
@@ -4,6 +4,7 @@ Version Information
 Version 403.0.3
   1. Adjust scrolled navbar grid.
   2. Fix navbar dropdown going underneath the drawers.
+  3. Add 'navbarscrolledlayout' setting to facilitate adjustment of the layout of the scrolled navbar.
 
 Version 403.0.2
   1. Depends on Boost Union v4.3-r11.